# docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenClaw Agent Marketplace API
  version: 0.1.0
servers:
  - url: http://localhost:4000
paths:
...
  /orders:
    get:
      summary: List orders
      tags: [Orders]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
    post:
      summary: Create order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
  /orders/{id}:
    get:
      summary: Order detail
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/Order"
                      - type: object
                        properties:
                          deliverables:
                            type: array
                            items:
                              $ref: "#/components/schemas/Deliverable"
        "404":
          $ref: "#/components/responses/NotFound"
  /orders/{id}/transition:
    post:
      summary: Advance order state
      tags: [Orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nextState:
                  $ref: "#/components/schemas/OrderState"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponse"
        "400":
          description: Invalid transition
...
components:
  responses:
    NotFound:
      description: Not Found
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
        buyerId:
          type: string
        templateId:
          type: string
        packageId:
          type: string
        inputs:
          type: object
        state:
          $ref: "#/components/schemas/OrderState"
        escrowAmount:
          type: number
        currency:
          type: string
        createdAt:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Order"
    OrderState:
      type: string
      enum:
        - created
        - accepted
        - in_progress
        - delivered
        - rejected
        - accepted_final
        - settled
        - cancelled
        - disputed
    Deliverable:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        version:
          type: integer
        artifactUrl:
          type: string
        summary:
          type: string
    CreateOrderInput:
      type: object
      required:
        - templateId
        - packageId
        - inputs
      properties:
        templateId:
          type: string
        packageId:
          type: string
        buyerId:
          type: string
        inputs:
          type: object
        currency:
          type: string
        escrowAmount:
          type: number
...